fix(codegen): handle array members from typedef'd C++ structs (Issue #342)
When passing an array member (like msg.buf) from an external C++ struct
to a function expecting an array parameter, the transpiler was incorrectly
adding '&' to the argument, causing a compilation error:

  error: cannot convert 'uint8_t (*)[8]' to 'const uint8_t*'

Root cause: The C++ Symbol Collector wasn't collecting struct field info
for typedef'd anonymous structs (typedef struct { ... } Name), so the
transpiler couldn't detect that the member was an array.

Changes:
- CppSymbolCollector: Detect anonymous class specifiers in typedefs and
  collect their fields using the typedef name
- Add test for array member to array parameter passing
- Update existing tests for improved enum/const member handling

Fixes #342

+++ tests/functions/array-struct-member-cpp.expected.c
@@ -27,8 +27,8 @@ uint32_t testArrayStructMembers(void) {
     sensors[0].status = STATUS_ON;
     sensors[0].value = 42;
     crc = process(crc, sensors[0].active);
-    crc = process(crc, sensors[0].sensorType);
-    crc = process(crc, sensors[0].status);
+    crc = process(crc, static_cast<uint8_t>(sensors[0].sensorType));
+    crc = process(crc, static_cast<uint8_t>(sensors[0].status));
     crc = process(crc, sensors[0].value);
     return crc;
 }
@@ -40,8 +40,8 @@ uint32_t testMultipleIndices(void) {
     sensors[2].status = STATUS_ERROR;
     sensors[3].value = 100;
     crc = process(crc, sensors[1].active);
-    crc = process(crc, sensors[1].sensorType);
-    crc = process(crc, sensors[2].status);
+    crc = process(crc, static_cast<uint8_t>(sensors[1].sensorType));
+    crc = process(crc, static_cast<uint8_t>(sensors[2].status));
     crc = process(crc, sensors[3].value);
     return crc;
 }
